get initial ws connection set up
