+++ package.json
@@ -49,7 +49,8 @@
     "topojson-client": "^3.0.0",
     "typeface-roboto": "0.0.54",
     "ws": "^7.4.2",
-    "xml2json": "^0.11.2"
+    "xml2json": "^0.11.2",
+    "xrpl-client": "^1.9.3"
   },
   "devDependencies": {
     "@babel/eslint-parser": "^7.14.7",
